requirements.txt
pygame
python-telegram-bot
pyTelegramBotAPI
ntplib
clock
python-dotenv